Create a supporting Net::NTLM::Client::Session class
